lang/ocaml/ocaml.mk: Also set MAKE_ENV for !DUNE_BUILD_PACKAGES
There is an if, and two build commands, which are mostly the same, but
with differences for [mysterious dune package stuff].  The
non-dune-package arm of the if didn't set the MAKE_ENV.  I'm calling
that a bug and adding it.

I've been running with this for a while, building various things, and
it seems ok.  Proposed on tech-pkg with no objections and one positive
comment.  If this breaks things for you, feel free to revert without
asking me, and then you get to be in charge of making ocaml work.