security/zoneminder: Convert to cmake/build.mk
This had slightly more changes than expected, but nothing mysterious:
manual install rules need to reference built scripts in the cmake
build directory.

pkglint indent level cleanup.  Note that pkglint is not perceiving
ZM_PERL_PATH as a list and incorrectly warning (but that is hard to
deal with).

Update list of known-working database versions, and leave a hint that
others have not been tested.

Build tested only.