pkgin: Add pkgin-static option.
This ensures pkgin links all pkgsrc libraries statically.  I've used a
variant of this patch for my illumos and macOS binary builds for a long
time now, and it works very well to eliminate a whole class of potential
upgrade problems.

Not enabled by default, as it can be specific to each OS and
configuration, but is highly recommended for anyone distributing binary
package repositories.