Fix kqueue use in libowfat.
Without this, libowfat does not handle kqueue read/write events on
file descriptors correctly.  A typical result of this is a consumer
spinning at 100% CPU instead of blocking until data is available for
reading.

Reported upstream and fixed there in CVS, so hopefully this patch
can be removed with the next release.