devel/py-mercurial: Disable rusty tests.
If we're not building the rusty parts of Mercurial, it doesn't make
sense for the tests to pick up whether $PREFIX/bin/cargo happens to
be installed to decide whether to run the rusty tests of Mercurial.

(If we do enable the rusty parts of Mercurial, whether via options or
unconditionally, we should run the rusty tests too under the same
conditions.)

No revbump because no change to the build -- only to tests.