libjpeg-turbo: updated to 3.1.0
3.1.0

libjpeg-turbo-3.1.0.tar.gz is the official source tarball for this release. The automatically generated "Source code" assets are not supported.
Refer to https://libjpeg-turbo.org/Downloads/DigitalSignatures for information regarding the methods used to sign the files in this release and instructions for verifying the signatures.
Refer to https://libjpeg-turbo.org/Documentation/OfficialBinaries for information regarding the compatibility of the binary packages in this release.

Packaging Changes

The RPM packages now contain SHA-256 header and payload digests. This fixes an issue whereby the RPM signatures could not be verified on Red Hat Enterprise Linux with FIPS mode enabled. The RPM packages now require GLIBC 2.17 or later.