gopls: update to 0.17.1.
0.17.1

This release fixes two crashes in gopls@v0.17.0:

golang/go#70889: a crash in completion of type instances inside a type conversion (found via telemetry).
golang/go#70927: a crash when a test file has a declaration with signature func(*error).

0.17.0

This release includes a variety of new features, bug fixes, and
performance improvements. It is also the first version of gopls to
require the latest released version of the Go toolchain, which
should be downloaded transparently during the gopls installation
process.