subversion: updated to 1.14.5
Apache Subversion 1.14.5. This is a stable bugfix and security release of the
Apache Subversion open source version control system. Among regular bug fixes,
this release fixes CVE-2024-46901: mod_dav_svn denial-of-service via control
characters in paths It has been discovered that the patch for CVE-2013-1968
was incomplete and unintentionally left mod_dav_svn vulnerable to control
characters in filenames. If a path or a revision-property which contains
control characters is committed to a repository then SVN operations served by
mod_dav_svn can be disrupted.