math/py-scpiy: c17 and add comments about compiler requirements
Declare that this needs c17, based on meson.build.  This has no effect
because it was marked for c++17 and both c17 and c++17 currently map
to gcc10.  Verified to still use gcc10 with the change.

Add comments about upstream requirements.

No PKGREVISION as this change should never have any effect on the
build.