py-pspdfutils: update to 3.3.8.
pkgsrc did not follow the rename below (yet?).

3.3.8

This release just removes some text accidentally added to the README
when the Python package name was changed in 3.3.7.

3.3.7

This release marks the long-awaited renaming on PyPI to “psutils”,
i.e. the same as the package name. Thanks to the Python Packaging
Authority for granting my application to use the name!

Also, some missing copyright notices are added to the command and
test modules, and the ‘extractres’ tests actually assert the results
of file comparisons.