lang/gcc10: Fix comment and nix space in LINK_LIBGCC_SPEC rpath flag.
Got this in all the other gcc versions but missed this one.  The
comment change is not particularly significant but let's make it
consistent between the parallel patches of gcc versions.

Revbump because this space change is significant on some platforms
and I don't know offhand whether it toggles between buildable or
not-buildable, or broken and working; in the latter case a revbump is
required.

PKGREVISION of gcc10 remains below PKGREVISION of gcc10-libs because
gcc10-libs got an extra revbump recently for the REQUIRES/PROVIDES
issue.

PR pkg/48740: lang/gcc44 to gcc46 linker fails because of whitespace