devel/extra-cmake-modules: Use buildlink3.mk, not TOOL_DEPENDS.
This one is borderline: should it be BUILD_DEPENDS or TOOL_DEPENDS?
It doesn't appear to vary depending on the build environment, so it
doesn't meaningfully describe the cross-builder vs the target system.
It is used only at build-time, but it doesn't have any native
executables -- they're just source code cmake scripts.

For now, it is convenient to use a buildlink3.mk file because this
enables us to explicitly expose only files under share/ECM/ -- and
not other random files installed in the prefix -- via buildlink3 to
keep dependent package builds clean.  Eventually we should have the
same mechanism for TOOL_DEPENDS but that day has not yet come.

This change creates a new use of BUILDLINK_DEPMETHOD build for
extra-cmake-modules, which is a priori risky.  Fortunately,
extra-cmake-modules has no libraries, so this is unlikely to cause
trouble.