fonts/harfbuzz: Bump BUILDLINK_ABI_DEPENDS to >=8.4.0.
This is necessary to detect whether introspection is included by
testing whether the `introspection' option is enabled.  If
introspection is included, we have to link against glib2, but prior
to 8.4.0 we could not detect this by testing the `introspection'
option because it was only introduced between our 8.3.0 and 8.4.0
packages (and I neglected to revbump when I added the option, so we
can't use 8.3.0nb1 or something -- has to be 8.4.0).