devel/py-boost: Need boost-libs to satisfy REQUIRES dependencies.
Resolves:

=> Checking for missing run-time search paths in py313-boost-1.87.0nb1
ERROR: lib/libboost_numpy313.so.1.87.0: missing library: libboost_container.so.1.87.0
ERROR: lib/libboost_numpy313.so.1.87.0: missing library: libboost_graph.so.1.87.0
ERROR: lib/libboost_python313.so.1.87.0: missing library: libboost_graph.so.1.87.0
ERROR: lib/libboost_python313.so.1.87.0: missing library: libboost_container.so.1.87.0
*** Error code 1

It's kind of kooky that we build boost-graph and boost-container as
part of the boost-python build, and then delete them, instead of
reusing them from boost-libs, but finding a way to do that will
require wading into the boost build system.  So let's just try to
make sure they're available for REQUIRES/PROVIDES analysis.

Adding this dependency is unlikely to be onerous -- applications
using boost-python are almost certainly going to need the rest of
boost anyway (and they already needed it for boost-graph and
boost-container!).