devel/hs-tasty-golden: import hs-tasty-golden-2.3.5
This package provides support for "golden testing". A golden test is an IO
action that writes its result to a file. To pass the test, this output file
should be identical to the corresponding "golden" file, which contains the
correct result for the test. To get started with golden testing and this
library, see Introduction to golden testing
(https://ro-che.info/articles/2017-12-04-golden-tests).