mk: Stop using 'sed -b' on Cygwin.
If this was previously added to work around an issue, it appears to no
longer be necessary, and was the cause of libtool build failures that
resulted in bootstrap being changed to use nbsed.  Unfortunately that
change itself broke libtool, but in a way that was only apparent when
actually using libtool.

From Jonathan Perkin <jperkin@smartos.org>
via drecklypkg commit 1e650715dafc0b9b64b3c9e8ba4885f171ea95d4