www/ruby-cgi: add package version 0.4.2
ruby-cgi contains newer version of cgi gem which is default gem of Ruby.

This package intends to handle these security problem of cgi gem in Ruby.

	CVE-2025-27219: Denial of Service in CGI::Cookie.parse
	CVE-2025-27220: ReDoS in CGI::Util#escapeElement
	CVE-2025-27221: userinfo leakage in URI#join, URI#merge and URI#+


CGI

CGI is a large class, providing several categories of methods, many of which
are mixed in from other modules.  Some of the documentation is in this
class, some in the modules CGI::QueryExtension and CGI::HtmlExtension.  See
CGI::Cookie for specific information on handling cookies, and cgi/session.rb
(CGI::Session) for information on sessions.

For queries, CGI provides methods to get at environmental variables,
parameters, cookies, and multipart request data.  For responses, CGI
provides methods for writing output and generating HTML.