hdf5: updated to 1.14.6
1.14.6

This is a maintenance release with some minor new changes:

Reverted a change to the library’s handling of UTF-8 file names. A change was made in the HDF5 1.14.4 release to address some issues with the library’s handling of code pages and file paths on Windows. This change introduced other issues with handling UTF-8 file names that caused breakage for software using the 1.14.4 and 1.14.5 releases of HDF5.

The change has been reverted for this release and will be addressed differently in a future release.

If you use 1.14.4 or 1.14.5 and are happy with how UTF-8 file names are handled in those versions, do not update to 1.14.6.
Added support for native, i.e., built-into-the HDF5 library, zlib-ng compression. CMake find package call also finds system zlib.
Modified the CMake logic for finding the libaec compression library, a drop-in replacement for szip, for a system-installed library version.
Removed the CMake module search find_package for the szip library.