Added www/php-matomo version 5.5.2
Matomo is the leading Free/Libre open analytics platform.

Matomo is a full-featured PHP MySQL software program that you
download and install on your own webserver.  At the end of the
five-minute installation process, you will be given a JavaScript
code.  Simply copy and paste this tag on websites you wish to track
and access your analytics reports in real-time.

Matomo aims to be a Free software alternative to Google Analytics
and is already used on more than 1,400,000 websites. Privacy is
built-in!